Force infra tests to run first.
